Update sun.py
Add monitored_conditions config item to specify which attributes to maintain. If not specified use previous set as default. For elevation and azimuth, add scan_interval config item (minimum one minute.) If elevation and/or azimuth are not specified they will not be added as attributes, but they will still be maintained anyway so that sun state can be determined.