Sun (#3)
Add monitored_conditions config item to specify which attributes to maintain. If not specified then original set will be used. For elevation and azimuth, add scan_interval (minimum of one minute.) If next_rising or next_setting are not specified they will not be added as attributes. However, they will always be maintained anyway so that sun's state can be determined/maintained.