Update life360.py
If Life360 provides an avatar for member then use that for the entity's picture. Given the way the device_tracker component works, this only applies when the device is first added to known_devices.yaml. If a member is already in your known_devices.yaml file, and you want to use this picture from Life360, either delete the member from known_devices.yaml, or temporarily comment it out and then merge the entries after a restart.