Update life360.py
Add new config item - times_as_local - to cause datetime attributes - last_seen & at_loc_since - to be in the local timezone as opposed to UTC. Default to False. Also, move the import of the life360 PyPI module into a try block in case it didn't get installed to handle that scenario better.